Register pprof as a plugin
